refactor app structure and clean assets
