+++ README.md
@@ -1,6 +1,6 @@
 # life-tools
 
-一个个人自用的生活工具站首页原型。当前阶段只做前端：首页视觉、响应式适配、工具入口和占位交互。
+一个面向个人日常管理的轻量生活工具台，集中放置资产、订阅、共享账本和后续常用工具入口。
 
 ## 技术栈
 
@@ -12,16 +12,23 @@
 - Mantine
 - @tabler/icons-react
 
-## 当前首页
+## 项目结构
+
+- `src/app`：应用入口、路由和 Mantine 主题。
+- `src/features/auth`：登录、注册、账户菜单和认证上下文。
+- `src/features/home`：首页、工具卡片、工具数据和功能预览。
+- `src/lib/supabase`：Supabase client 和内部用户名映射。
+
+## 产品入口
 
 - 顶部：`life-tools` logo + 头像入口，不包含登录按钮。
 - 账户入口：右上角头像支持 Supabase 用户名密码登录 / 注册 / 退出。
-- 主视觉：标题和副标题，顶部区域保持紧凑。
-- 工具入口：资产管理、订阅管理、共享账本、更多工具。
+- 主视觉：用简洁标题说明产品定位，顶部区域保持紧凑。
+- 工具入口：资产台账、订阅日历、共享账本、工具库。
 - 桌面端：工具卡为横向长方形入口，宽屏 3 + 1 排布，第二行从左开始；中等宽度自动转 2 列。
 - 移动端：工具卡 2 列布局，保持紧凑卡片比例，压缩 header、标题区和卡片间距。
-- 占位交互：点击工具卡打开 Mantine Modal。
-- 登录交互：点击头像打开 Mantine Modal，只填写用户名 + 密码；底层使用 Supabase Auth。
+- 工具交互：点击工具卡打开功能预览，后续接入真实数据和操作流程。
+- 账户交互：点击头像打开账号菜单或登录弹窗，只填写用户名 + 密码；底层使用 Supabase Auth。
 
 ## Supabase 登录
 
@@ -40,7 +47,7 @@
 - 图标使用 `@tabler/icons-react`。
 - Logo 拆成 icon-only SVG + HTML 文本字标：`life-tools-logo.svg` 只包含房子图标，`life-tools` 由 CSS 控制字号、字重和间距，避免 SVG 字体/裁切问题。
 - 图标背景不使用切图，使用 CSS 渐变、内阴影和轻边框实现；icon tile 已调成更方正的圆角矩形，桌面卡片为图标左、文字中、箭头右下的横向结构。
-- 参考视觉稿保存在：`design-reference-home-v2.png`。
+- 阶段记录放在 `docs/STATUS.md`，设计验收记录放在 `docs/design-qa.md`。
 
 ## 本地开发
