+++ index.html
@@ -5,9 +5,11 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <meta
       name="description"
-      content="life-tools 是一个温暖、简洁、轻量的个人生活工具入口。"
+      content="life-tools 是一个用于管理资产、订阅和共享账本的个人生活工具台。"
     />
-    <title>life-tools</title>
+    <meta name="theme-color" content="#fbf7ef" />
+    <link rel="icon" type="image/svg+xml" href="/favicon.svg" />
+    <title>life-tools | 个人生活工具台</title>
   </head>
   <body>
     <div id="root"></div>
