+++ src/app/App.tsx
@@ -1,0 +1,20 @@
+import { MantineProvider } from "@mantine/core";
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import { RouterProvider } from "@tanstack/react-router";
+import { AuthProvider } from "../features/auth/auth-context";
+import { router } from "./router";
+import { theme } from "./theme";
+
+const queryClient = new QueryClient();
+
+export function App() {
+  return (
+    <MantineProvider theme={theme}>
+      <QueryClientProvider client={queryClient}>
+        <AuthProvider>
+          <RouterProvider router={router} />
+        </AuthProvider>
+      </QueryClientProvider>
+    </MantineProvider>
+  );
+}
